Remove ENTRYPOINT from Dockerfile.
Doing this because CircleCI doesn't use it properly anyway, so my
existing configuration already explicitly calls the script.

CircleCI seems to have made a change somewhere that is making the
`latest` tag not start right, but I don't want to change the default
Docker behaviour of that tag.

For that reason, I'm pushing this to a `no-entrypoint` tag, that I can
ensure works in CircleCI now.